+++ app/Views/layouts/template.php
@@ -86,41 +86,37 @@
 	<script>
 		var baseUrl = "<?= base_url();?>";
 		var siteUrl = "<?= site_url();?>";
+		
+		<?php if (isset($charts)) : ?>
+			//var charts = JSON.parse(`json_encode($charts)`);
+			var charts = <?=json_encode($charts) ?>;		
+		<?php endif; ?>		 
 	</script>
 	<script>
 var ctx = document.getElementById('myChart').getContext('2d');
 var myChart = new Chart(ctx, {
-    type: 'polarArea',
+    type: 'pie',
     data: {
-        labels: ['Red', 'Blue', 'Yellow', 'Green', 'Purple', 'Orange'],
+        labels: charts.label,
         datasets: [{
             label: '# of Votes',
-            data: [12, 19, 3, 5, 2, 3],
+            data: charts.data,
             backgroundColor: [
-                'rgba(255, 99, 132, 0.2)',
-                'rgba(54, 162, 235, 0.2)',
-                'rgba(255, 206, 86, 0.2)',
-                'rgba(75, 192, 192, 0.2)',
-                'rgba(153, 102, 255, 0.2)',
-                'rgba(255, 159, 64, 0.2)'
+                '#29cc97',
+                '#0fa0b2',
+                '#fec400',
+                '#fe5461',
             ],
             borderColor: [
-                'rgba(255, 99, 132, 1)',
-                'rgba(54, 162, 235, 1)',
-                'rgba(255, 206, 86, 1)',
-                'rgba(75, 192, 192, 1)',
-                'rgba(153, 102, 255, 1)',
-                'rgba(255, 159, 64, 1)'
+                '#29cc97',
+                '#0fa0b2',
+                '#fec400',
+                '#fe5461',
             ],
             borderWidth: 1
         }]
     },
     options: {
-        scales: {
-            y: {
-                beginAtZero: true
-            }
-        }
     }
 });
 </script>